+++ crates/sui-types/src/layout_resolver.rs
@@ -38,7 +38,7 @@ pub fn get_layout_from_struct_tag(
 
 pub fn into_struct_layout(layout: A::MoveDatatypeLayout) -> Result<A::MoveStructLayout, SuiError> {
     match layout {
-        A::MoveDatatypeLayout::Struct(s) => Ok(s),
+        A::MoveDatatypeLayout::Struct(s) => Ok(*s),
         A::MoveDatatypeLayout::Enum(e) => Err(SuiError::ObjectSerializationError {
             error: format!("Expected struct layout but got an enum {e:?}"),
         }),