+++ crates/suiop-cli/src/cli/incidents/mod.rs
@@ -61,21 +61,7 @@ pub async fn incidents_cmd(args: &IncidentsArgs) -> Result<()> {
 
             let incidents = fetch_incidents(*limit, start_time, current_time).await?;
             if *interactive {
-                println!("{:?} incidents found", incidents);
-                review_recent_incidents(
-                    incidents
-                        .into_iter()
-                        // filter on priority > P3 and any slack channel association
-                        .filter(|i| {
-                            i.priority
-                                .clone()
-                                .filter(|p| !p.name.is_empty() && p.name != "P3")
-                                .is_some()
-                                || i.slack_channel.is_some()
-                        })
-                        .collect::<Vec<_>>(),
-                )
-                .await?
+                review_recent_incidents(incidents).await?
             } else {
                 print_recent_incidents(incidents, *long, *with_priority).await?
             }