+++ external-crates/move/crates/move-compiler/src/linters/mod.rs
@@ -13,10 +13,11 @@ use crate::{
 
 pub mod abort_constant;
 pub mod constant_naming;
+pub mod loop_without_exit;
 pub mod meaningless_math_operation;
+pub mod redundant_conditional;
 pub mod unnecessary_while_loop;
 pub mod unneeded_return;
-pub mod redundant_conditional;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum LintLevel {
@@ -104,7 +105,7 @@ lints!(
     ),
     (
         WhileTrueToLoop,
-        LinterDiagnosticCategory::Complexity,
+        LinterDiagnosticCategory::Style,
         "while_true",
         "unnecessary 'while (true)', replace with 'loop'"
     ),
@@ -125,6 +126,12 @@ lints!(
         LinterDiagnosticCategory::Style,
         "abort_without_constant",
         "'abort' or 'assert' without named constant"
+    ),
+    (
+        LoopWithoutExit,
+        LinterDiagnosticCategory::Suspicious,
+        "loop_without_exit",
+        "'loop' without 'break' or 'return'"
     ),
     (
         RedundantConditional,
@@ -164,6 +171,7 @@ pub fn linter_visitors(level: LintLevel) -> Vec<Visitor> {
                 meaningless_math_operation::MeaninglessMathOperation.visitor(),
                 unneeded_return::UnneededReturnVisitor.visitor(),
                 abort_constant::AssertAbortNamedConstants.visitor(),
+                loop_without_exit::LoopWithoutExit.visitor(),
                 redundant_conditional::RedundantConditional.visitor(),
             ]
         }