+++ external-crates/move/crates/move-vm-config/src/runtime.rs
@@ -36,6 +36,9 @@ pub struct VMConfig {
     // Whether value serialization errors when generating type layouts should be rethrown or
     // converted to a different error.
     pub rethrow_serialization_type_layout_errors: bool,
+    /// Maximal nodes which are allowed when converting to layout. This includes the types of
+    /// fields for struct types.
+    pub max_type_to_layout_nodes: Option<u64>,
 }
 
 impl Default for VMConfig {
@@ -50,6 +53,7 @@ impl Default for VMConfig {
             error_execution_state: true,
             binary_config: BinaryConfig::with_extraneous_bytes_check(false),
             rethrow_serialization_type_layout_errors: false,
+            max_type_to_layout_nodes: Some(512),
         }
     }
 }