+++ external-crates/move/crates/move-vm-runtime/src/runtime.rs
@@ -9,6 +9,7 @@ use crate::{
     native_extensions::NativeContextExtensions,
     native_functions::{NativeFunction, NativeFunctions},
     session::{LoadedFunctionInstantiation, SerializedReturnValues, Session},
+    tracing2::tracer::VMTracer,
 };
 use move_binary_format::{
     errors::{verification_error, Location, PartialVMError, PartialVMResult, VMResult},
@@ -25,6 +26,7 @@ use move_core_types::{
     runtime_value::MoveTypeLayout,
     vm_status::StatusCode,
 };
+use move_trace_format::format::MoveTraceBuilder;
 use move_vm_config::runtime::VMConfig;
 use move_vm_types::{
     data_store::DataStore,
@@ -320,6 +322,7 @@ impl VMRuntime {
         data_store: &mut impl DataStore,
         gas_meter: &mut impl GasMeter,
         extensions: &mut NativeContextExtensions,
+        tracer: &mut Option<VMTracer<'_>>,
     ) -> VMResult<SerializedReturnValues> {
         let arg_types = param_types
             .into_iter()
@@ -351,6 +354,7 @@ impl VMRuntime {
             gas_meter,
             extensions,
             &self.loader,
+            tracer,
         )?;
 
         let serialized_return_values = self
@@ -391,6 +395,7 @@ impl VMRuntime {
         gas_meter: &mut impl GasMeter,
         extensions: &mut NativeContextExtensions,
         bypass_declared_entry_check: bool,
+        tracer: Option<&mut MoveTraceBuilder>,
     ) -> VMResult<SerializedReturnValues> {
         use move_binary_format::file_format::SignatureIndex;
         fn check_is_entry(
@@ -442,6 +447,7 @@ impl VMRuntime {
             data_store,
             gas_meter,
             extensions,
+            &mut tracer.map(VMTracer::new),
         )
     }
 
@@ -511,6 +517,7 @@ impl VMRuntime {
             gas_meter,
             extensions,
             bypass_declared_entry_check,
+            None,
         )
     }